+++ src/Text/Show/Text/Data/Monoid.hs
@@ -1,0 +1,9 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+module Text.Show.Text.Data.Monoid where
+
+import Data.Text.Lazy.Builder (Builder)
+
+import Prelude hiding (Show)
+
+import Text.Show.Text.Class (Show)