+++ src/Text/Show/Text/Functions.hs
@@ -1,4 +1,5 @@
-{-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE NoImplicitPrelude, OverloadedStrings #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Text.Show.Text.Functions
@@ -8,43 +9,18 @@
 -- Stability   :  Experimental
 -- Portability :  GHC
 -- 
--- Miscellaneous 'Builder' utility functions.
+-- Optional 'Show' instance for functions.
 ----------------------------------------------------------------------------
-module Text.Show.Text.Functions where
+module Text.Show.Text.Functions (showbFunction) where
 
-import Data.Int (Int64)
-import Data.Monoid (Monoid(mempty), (<>))
-import Data.Text.Lazy (length, replicate)
-import Data.Text.Lazy.Builder (Builder, fromLazyText, singleton, toLazyText)
+import Data.Text.Lazy.Builder (Builder)
+import Prelude hiding (Show)
+import Text.Show.Text.Class (Show(showb))
 
-import Prelude hiding (length, replicate)
+showbFunction :: (a -> b) -> Builder
+showbFunction = const "<function>"
+{-# INLINE showbFunction #-}
 
--- |
--- A shorter name for 'singleton' for convenience's sake (since it tends to be used
--- pretty often in @text-show@).
-s :: Char -> Builder
-s = singleton
-{-# INLINE s #-}
-
--- | Computes the length of a 'Builder'.
-lengthB :: Builder -> Int64
-lengthB = length . toLazyText
-{-# INLINE lengthB #-}
-
--- | @'replicateB' n b@ yields a 'Builder' containing @b@ repeated @n@ times.
-replicateB :: Int64 -> Builder -> Builder
-replicateB n = fromLazyText . replicate n . toLazyText
-{-# INLINE replicateB #-}
-
--- | Merges several 'Builder's, separating them by newlines.
-unlinesB :: [Builder] -> Builder
-unlinesB (b:bs) = b <> s '\n' <> unlinesB bs
-unlinesB []     = mempty
-{-# INLINE unlinesB #-}
-
--- | Merges several 'Builder's, separating them by spaces.
-unwordsB :: [Builder] -> Builder
-unwordsB (b:bs@(_:_)) = b <> s ' ' <> unwordsB bs
-unwordsB [b]          = b
-unwordsB []           = mempty
-{-# INLINE unwordsB #-}
+instance Show (a -> b) where
+    showb = showbFunction
+    {-# INLINE showb #-}