+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Show instances for many other data types in `base`, `containers` and `time`
 * Expose internal modules with monomorphic functions
 * `Text.Show.Text` now exports `Data.Text.Lazy.Builder` for convenience
-* Add `showLazy`, `printLazy`, `lengthB`, and `replicateB`
+* Add `showLazy`, `printLazy`, `hPrint`, `hPrintLazy`, `lengthB`, and `replicateB`
 
 # 0.2.0.0